requirements.txt
streamlit
google-generativeai
rank_bm25
sentence-transformers
numpy
